# test-api-server/openapi.yaml
openapi: 3.0.3
info:
  title: MCP Test API Server
  description: |
    A comprehensive API server for testing MCP Studio functionality. 
    Provides CRUD operations for users, chats, messages, and feedbacks with MongoDB backend.

    This API is designed to be easily convertible to an MCP (Model Context Protocol) server.
  version: 1.0.0
  contact:
    name: MCP Studio Team
    email: support@mcpstudio.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.mcpstudio.com
    description: Production server
...
paths:
...
  /api/chats:
    get:
      tags: [chats]
      summary: Get all chats
      description: Retrieve a paginated list of all chats
      operationId: getChats
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Chats retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatListResponse"
        "500":
          description: Failed to fetch chats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      tags: [chats]
      summary: Create new chat
      description: Create a new chat for a user
      operationId: createChat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChatRequest"
      responses:
        "201":
          description: Chat created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateChatResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Chat already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Failed to create chat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/chats/{id}:
    get:
      tags: [chats]
      summary: Get chat by ID
      description: Retrieve a specific chat by its ID
      operationId: getChatById
      parameters:
        - name: id
          in: path
          required: true
          description: Chat ID
          schema:
            type: string
      responses:
        "200":
          description: Chat retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"
        "404":
          description: Chat not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Failed to fetch chat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      tags: [chats]
      summary: Delete chat
      description: Delete a chat and all related messages and feedbacks
      operationId: deleteChat
      parameters:
        - name: id
          in: path
          required: true
          description: Chat ID
          schema:
            type: string
      responses:
        "200":
          description: Chat deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteChatResponse"
        "404":
          description: Chat not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Failed to delete chat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/chats/user/{userId}:
    get:
      tags: [chats]
      summary: Get chats by user ID
      description: Retrieve all chats owned by a specific user
      operationId: getChatsByUserId
      parameters:
        - name: userId
          in: path
          required: true
          description: User ID (ObjectId)
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: User chats retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatListResponse"
        "400":
          description: Invalid user ID format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Failed to fetch user chats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/messages/chat/{chatId}:
    get:
      tags: [messages]
      summary: Get messages by chat ID
      description: Retrieve all messages for a specific chat
      operationId: getMessagesByChatId
      parameters:
        - name: chatId
          in: path
          required: true
          description: Chat ID
          schema:
            type: string
        - $ref: "#/components/parameters/Page"
        - name: limit
          in: query
          description: Number of messages per page (default 50)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: Messages retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageListResponse"
        "500":
          description: Failed to fetch messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      tags: [messages]
      summary: Delete all messages for chat
      description: Delete all messages in a specific chat
      operationId: deleteMessagesByChatId
      parameters:
        - name: chatId
          in: path
          required: true
          description: Chat ID
          schema:
            type: string
      responses:
        "200":
          description: Messages deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "500":
          description: Failed to delete messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # Feedback endpoints
...
  /api/feedbacks/chat/{chatId}:
    get:
      tags: [feedbacks]
      summary: Get feedbacks by chat ID
      description: Retrieve all feedbacks for a specific chat
      operationId: getFeedbacksByChatId
      parameters:
        - name: chatId
          in: path
          required: true
          description: Chat ID
          schema:
            type: string
        - $ref: "#/components/parameters/Page"
        - name: limit
          in: query
          description: Number of feedbacks per page (default 20)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Feedbacks retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackListResponse"
        "500":
          description: Failed to fetch feedbacks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      tags: [feedbacks]
      summary: Delete all feedbacks for chat
      description: Delete all feedbacks in a specific chat
      operationId: deleteFeedbacksByChatId
      parameters:
        - name: chatId
          in: path
          required: true
          description: Chat ID
          schema:
            type: string
      responses:
        "200":
          description: Feedbacks deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "500":
          description: Failed to delete feedbacks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  parameters:
    Page:
      name: page
      in: query
      description: Page number for pagination (optional - only used when both page and limit are provided)
      schema:
        type: integer
        minimum: 1
    Limit:
      name: limit
      in: query
      description: Number of items per page (optional - only used when both page and limit are provided, max 100)
      schema:
        type: integer
        minimum: 1
        maximum: 100
  schemas:
...
    ObjectId:
      type: string
      pattern: "^[0-9a-fA-F]{24}$"
      description: MongoDB ObjectId
    Timestamp:
      type: string
      format: date-time
      description: ISO 8601 timestamp

    # Error responses
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Detailed error message

    # Pagination (optional - only when page and limit are both provided)
    PaginationOptional:
      type: object
      properties:
        current:
          type: integer
          description: Current page number
        total:
          type: integer
          description: Total number of pages
        count:
          type: integer
          description: Number of items in current page
        totalRecords:
          type: integer
          description: Total number of records

    # Base list response without pagination
...
    SharedUser:
      type: object
      required:
        - userId
        - name
        - email
        - addedAt
      properties:
        userId:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        addedAt:
          $ref: "#/components/schemas/Timestamp"
    Chat:
      type: object
      required:
        - _id
        - id
        - title
        - userId
        - createdAt
        - lastModifiedAt
        - visibility
        - sharedWith
      properties:
        _id:
          $ref: "#/components/schemas/ObjectId"
        id:
          type: string
          description: Custom chat ID
        title:
          type: string
        userId:
          $ref: "#/components/schemas/ObjectId"
        createdAt:
          $ref: "#/components/schemas/Timestamp"
        lastModifiedAt:
          $ref: "#/components/schemas/Timestamp"
        visibility:
          type: string
          enum: [private, shared]
        sharedWith:
          type: array
          items:
            $ref: "#/components/schemas/SharedUser"
    CreateChatRequest:
      type: object
      required:
        - id
        - userId
        - title
      properties:
        id:
          type: string
          description: Custom chat ID
        userId:
          $ref: "#/components/schemas/ObjectId"
        title:
          type: string
...
    CreateChatResponse:
      type: object
      required:
        - message
        - chat
      properties:
        message:
          type: string
        chat:
          $ref: "#/components/schemas/Chat"
...
    ChatListResponse:
      type: object
      required:
        - chats
        - totalRecords
      properties:
        chats:
          type: array
          items:
            $ref: "#/components/schemas/Chat"
        totalRecords:
          type: integer
          description: Total number of chats in the collection
        pagination:
          $ref: "#/components/schemas/PaginationOptional"
...
    DeleteChatResponse:
      type: object
      required:
        - message
        - deletedCounts
      properties:
        message:
          type: string
        deletedCounts:
          type: object
          required:
            - chats
            - messages
            - feedbacks
          properties:
            chats:
              type: integer
            messages:
              type: integer
            feedbacks:
              type: integer

    # Message schemas
    Message:
      type: object
      required:
        - _id
        - chatId
        - content
        - createdAt
        - updatedAt
      properties:
        _id:
          $ref: "#/components/schemas/ObjectId"
        chatId:
          type: string
        content:
          type: string
        role:
          type: string
          enum: [user, assistant, system]
        metadata:
          type: object
          additionalProperties: true
        createdAt:
          $ref: "#/components/schemas/Timestamp"
        updatedAt:
          $ref: "#/components/schemas/Timestamp"
...
    MessageListResponse:
      type: object
      required:
        - messages
        - totalRecords
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        totalRecords:
          type: integer
          description: Total number of messages in the collection
        pagination:
          $ref: "#/components/schemas/PaginationOptional"

    # Feedback schemas
    JiraIssue:
      type: object
      required:
        - issueKey
        - issueId
        - issueUrl
        - summary
        - description
        - status
        - priority
        - labels
        - createdDate
      properties:
        issueKey:
          type: string
        issueId:
          type: string
        issueUrl:
          type: string
        summary:
          type: string
        description:
          type: string
        status:
          type: string
        assignee:
          type: string
        priority:
          type: string
        labels:
          type: array
          items:
            type: string
        attachments:
          type: array
          items:
            type: string
        createdDate:
          $ref: "#/components/schemas/Timestamp"
    Feedback:
      type: object
      required:
        - _id
        - chatId
        - messageId
        - userId
        - comments
        - rating
        - isUpvoted
        - hasJiraIssue
        - createdAt
        - lastModifiedAt
      properties:
        _id:
          $ref: "#/components/schemas/ObjectId"
        chatId:
          type: string
        messageId:
          type: string
        userId:
          $ref: "#/components/schemas/ObjectId"
        comments:
          type: string
        rating:
          type: integer
          minimum: 0
          maximum: 5
        isUpvoted:
          type: boolean
        hasJiraIssue:
          type: boolean
        jiraIssue:
          $ref: "#/components/schemas/JiraIssue"
        createdAt:
          $ref: "#/components/schemas/Timestamp"
        lastModifiedAt:
          $ref: "#/components/schemas/Timestamp"
...
    FeedbackListResponse:
      type: object
      required:
        - feedbacks
        - totalRecords
      properties:
        feedbacks:
          type: array
          items:
            $ref: "#/components/schemas/Feedback"
        totalRecords:
          type: integer
          description: Total number of feedbacks in the collection
        pagination:
          $ref: "#/components/schemas/PaginationOptional"
...
    DeleteResponse:
      type: object
      required:
        - message
        - deletedCount
      properties:
        message:
          type: string
        deletedCount:
          type: integer
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
security:
  - bearerAuth: []